organization: move each oracle scripts into their own directories
